+++ source/source_pw/module_dfpt/dfpt_pert.cpp
@@ -1,18 +1,32 @@
+// ============================================================
+// This code is added by Mohan Chen on 2026-05-18.
+// This code is currently in the design phase and has not been
+// put into production yet. It may change in the future.
+// Please use this code with caution. Only developers who know
+// what they are doing should use this code.
+// ============================================================
+//
+// The KB-projector construction of DFPT_Pert (radial_vq, real_ylm,
+// grad_real_ylm, build_vkb, build_vkb_dk) lives in dfpt_pert_vkb.cpp
+// and the nonlocal first/second-order potentials (dVnl_dtau,
+// apply_d2vnl) in dfpt_pert_nl.cpp.
+
 #include "dfpt_pert.h"
 
 #include "source_base/constants.h"
 #include "source_base/global_function.h"
 #include "source_base/math_integral.h"
-#include "source_base/math_sphbes.h"
 #include "source_base/truncated_func.h"
 #include "source_cell/atom_pseudo.h"
 #include "source_cell/atom_spec.h"
 #include "source_pw/module_pwdft/stru_fac.h"
 
 #include <cmath>
+#include <complex>
 #include <cstdlib>
 #include <fstream>
 #include <string>
+#include <vector>
 
 #include "source_base/timer.h"
 #include "source_base/tool_title.h"
@@ -337,465 +351,6 @@ void DFPT_Pert::apply_dv(int q_idx, int k_idx, const psi::Psi<std::complex<doubl
     ModuleBase::timer::end("DFPT_Pert", "apply_dv");
 }
 
-// ---------------------------------------------------------------------------
-// nonlocal first-order potential (normal-conserving separable case)
-// ---------------------------------------------------------------------------
-
-double DFPT_Pert::radial_vq(int it, int ib, double g) const
-{
-    ModuleBase::TITLE("DFPT_Pert", "radial_vq");
-    ModuleBase::timer::start("DFPT_Pert", "radial_vq");
-    const pseudo& ncpp = ucell_->atoms[it].ncpp;
-    const int l = ncpp.lll[ib];
-    int kkbeta = ncpp.kkbeta;
-    if (kkbeta > 0 && (kkbeta % 2 == 0))
-    {
-        --kkbeta;
-    }
-    std::vector<double> jl(kkbeta);
-    std::vector<double> aux(kkbeta);
-    ModuleBase::Sphbes::Spherical_Bessel(kkbeta, ncpp.r.data(), g, l, jl.data());
-    for (int ir = 0; ir < kkbeta; ++ir)
-    {
-        aux[ir] = ncpp.betar(ib, ir) * jl[ir] * ncpp.r[ir];
-    }
-    double v = 0.0;
-    ModuleBase::Integral::Simpson_Integral(kkbeta, aux.data(), ncpp.rab.data(), v);
-    // tab convention from vnl_pw.cpp: (4pi/sqrt(Omega)) * integral
-    ModuleBase::timer::end("DFPT_Pert", "radial_vq");
-    return v * ModuleBase::FOUR_PI / std::sqrt(ucell_->omega);
-}
-
-double DFPT_Pert::real_ylm(int l, int m, const ModuleBase::Vector3<double>& ghat) const
-{
-    ModuleBase::TITLE("DFPT_Pert", "real_ylm");
-    ModuleBase::timer::start("DFPT_Pert", "real_ylm");
-    // orthonormal real spherical harmonics Y_{l,m} for l <= 2 with the
-    // standard convention, m in [-l, l]:
-    //   Y_{l,0}      = sqrt((2l+1)/4pi) P_l^0(cos0)
-    //   Y_{l,m>0}    = sqrt(2 (2l+1)/4pi (l-m)!/(l+m)!) P_l^m(cos0) cos(m phi)
-    //   Y_{l,m<0}    = sqrt(2 (2l+1)/4pi (l-|m|)!/(l+|m|)!) P_l^{|m|}(cos0) sin(|m| phi)
-    // with the associated Legendre convention P_1^1 = -sin0, P_2^1 = -3 sin0 cos0,
-    // P_2^2 = 3 sin^2 0. The ABACUS GS vkb applies an additional (-1)^|m| phase
-    // for the m>0 channels; exact GS parity is reconciled in the diamond
-    // end-to-end test (C7), while the C1 identity test is convention-independent.
-    const double ghat_zero_tol = 1.0e-12; ///< empirical parameter: |ghat| floor for the zero-direction limit
-    const double x = ghat.x;
-    const double y = ghat.y;
-    const double z = ghat.z;
-    const double r = std::sqrt(x * x + y * y + z * z);
-    if (r < ghat_zero_tol)
-    {
-        ModuleBase::timer::end("DFPT_Pert", "real_ylm");
-        return (l == 0) ? 0.5 * std::sqrt(1.0 / ModuleBase::PI) : 0.0;
-    }
-    const double nx = x / r;
-    const double ny = y / r;
-    const double nz = z / r;
-    switch (l)
-    {
-    case 0: {
-        ModuleBase::timer::end("DFPT_Pert", "real_ylm");
-        return 0.5 * std::sqrt(1.0 / ModuleBase::PI);
-    }
-    case 1: {
-        switch (m)
-        {
-        case -1:
-            ModuleBase::timer::end("DFPT_Pert", "real_ylm");
-            return -0.5 * std::sqrt(3.0 / ModuleBase::PI) * ny;
-        case 0:
-            ModuleBase::timer::end("DFPT_Pert", "real_ylm");
-            return 0.5 * std::sqrt(3.0 / ModuleBase::PI) * nz;
-        case 1:
-            ModuleBase::timer::end("DFPT_Pert", "real_ylm");
-            return -0.5 * std::sqrt(3.0 / ModuleBase::PI) * nx;
-        }
-        break;
-    }
-    case 2: {
-        switch (m)
-        {
-        case -2:
-            ModuleBase::timer::end("DFPT_Pert", "real_ylm");
-            return 0.5 * std::sqrt(15.0 / ModuleBase::PI) * nx * ny;
-        case -1:
-            ModuleBase::timer::end("DFPT_Pert", "real_ylm");
-            return -0.5 * std::sqrt(15.0 / ModuleBase::PI) * nz * ny;
-        case 0:
-            ModuleBase::timer::end("DFPT_Pert", "real_ylm");
-            return 0.25 * std::sqrt(5.0 / ModuleBase::PI) * (3.0 * nz * nz - 1.0);
-        case 1:
-            ModuleBase::timer::end("DFPT_Pert", "real_ylm");
-            return -0.5 * std::sqrt(15.0 / ModuleBase::PI) * nz * nx;
-        case 2:
-            ModuleBase::timer::end("DFPT_Pert", "real_ylm");
-            return 0.25 * std::sqrt(15.0 / ModuleBase::PI) * (nx * nx - ny * ny);
-        }
-        break;
-    }
-    default: {
-        ModuleBase::WARNING_QUIT("DFPT_Pert::real_ylm", "real_ylm implemented for l<=2 only (DFPT NC path).");
-    }
-    }
-    ModuleBase::timer::end("DFPT_Pert", "real_ylm");
-    return 0.0;
-}
-
-void DFPT_Pert::build_vkb(int it,
-                          int ia,
-                          const std::vector<ModuleBase::Vector3<double>>& gk,
-                          std::vector<std::vector<std::complex<double>>>& vkb) const
-{
-    ModuleBase::TITLE("DFPT_Pert", "build_vkb");
-    ModuleBase::timer::start("DFPT_Pert", "build_vkb");
-    // per-type projector bookkeeping mirrors the ground-state vnl_pw.cpp layout:
-    // every radial beta (nbeta) with angular momentum l spins out (2l+1)
-    // projectors with combined index lm = l^2 + m, m in 0..2l (i.e. the real
-    // harmonic m channels -l..l walked as m' = (-1)^(m+1) ceil... ABACUS ylm
-    // block: m=0, +1, -1, +2, -2, ...). We use the signed m' directly.
-    const pseudo& ncpp = ucell_->atoms[it].ncpp;
-    const int nh = ncpp.nh;
-    const int ngk = static_cast<int>(gk.size());
-    const ModuleBase::Vector3<double>& tau = ucell_->atoms[it].tau[ia];
-    vkb.assign(nh, std::vector<std::complex<double>>(ngk, std::complex<double>(0.0, 0.0)));
-    if (nh == 0)
-    {
-        ModuleBase::timer::end("DFPT_Pert", "build_vkb");
-        return;
-    }
-    int mu = 0;
-    for (int ib = 0; ib < ncpp.nbeta; ++ib)
-    {
-        const int l = ncpp.lll[ib];
-        if (l > 2)
-        {
-            ModuleBase::WARNING_QUIT("DFPT_Pert::build_vkb", "DFPT NC projector path implemented for l<=2 only.");
-        }
-        const std::complex<double> pref = std::pow(std::complex<double>(0.0, -1.0), l); // (-i)^l
-        for (int m = 0; m < 2 * l + 1; ++m)
-        {
-            // ABACUS real-harmonic walk over the m channels of this radial beta:
-            // m=0 -> m'=0; m=1 -> m'=+1; m=2 -> m'=-1; m=3 -> m'=+2; m=4 -> m'=-2
-            const int mr = (m == 0) ? 0 : ((m % 2 == 1) ? (m + 1) / 2 : -(m / 2));
-            for (int ig = 0; ig < ngk; ++ig)
-            {
-                const ModuleBase::Vector3<double>& G = gk[ig];         // k(+q)+G, 2*pi/lat0
-                const double gnorm = std::sqrt(G * G) * ucell_->tpiba; // bohr^-1
-                // real_ylm handles the |G|=0 point itself (Y_00 is
-                // direction-independent; l>0 channels vanish there together
-                // with vq), so the raw vector is passed directly.
-                const double ylm = real_ylm(l, mr, G);
-                const double vq = radial_vq(it, ib, gnorm);
-                // GS structure-factor convention (stru_fac.cpp get_sk /
-                // eigts, ci_tpi = -2pi i): exp(-i 2pi (gk.tau))
-                const double arg = -ModuleBase::TWO_PI * (G * tau);
-                const std::complex<double> phase(std::cos(arg), std::sin(arg));
-                vkb[mu][ig] = pref * ylm * vq * phase;
-            }
-            ++mu;
-        }
-    }
-    ModuleBase::timer::end("DFPT_Pert", "build_vkb");
-}
-
-void DFPT_Pert::grad_real_ylm(int l, int m, const ModuleBase::Vector3<double>& ghat, double grad[3]) const
-{
-    ModuleBase::TITLE("DFPT_Pert", "grad_real_ylm");
-    ModuleBase::timer::start("DFPT_Pert", "grad_real_ylm");
-    // analytic gradients of the real_ylm polynomials (l <= 2), consistent
-    // with the conventions documented above real_ylm
-    const double x = ghat.x;
-    const double y = ghat.y;
-    const double z = ghat.z;
-    const double c1 = 0.5 * std::sqrt(3.0 / ModuleBase::PI);
-    const double c2 = 0.5 * std::sqrt(15.0 / ModuleBase::PI);
-    const double c20 = 0.25 * std::sqrt(5.0 / ModuleBase::PI);
-    grad[0] = grad[1] = grad[2] = 0.0;
-    switch (l)
-    {
-    case 0:
-        ModuleBase::timer::end("DFPT_Pert", "grad_real_ylm");
-        return;
-    case 1:
-        switch (m)
-        {
-        case -1:
-            grad[1] = -c1;
-            ModuleBase::timer::end("DFPT_Pert", "grad_real_ylm");
-            return;
-        case 0:
-            grad[2] = c1;
-            ModuleBase::timer::end("DFPT_Pert", "grad_real_ylm");
-            return;
-        case 1:
-            grad[0] = -c1;
-            ModuleBase::timer::end("DFPT_Pert", "grad_real_ylm");
-            return;
-        }
-        break;
-    case 2:
-        switch (m)
-        {
-        case -2:
-            grad[0] = c2 * y;
-            grad[1] = c2 * x;
-            ModuleBase::timer::end("DFPT_Pert", "grad_real_ylm");
-            return;
-        case -1:
-            grad[1] = -c2 * z;
-            grad[2] = -c2 * y;
-            ModuleBase::timer::end("DFPT_Pert", "grad_real_ylm");
-            return;
-        case 0:
-            grad[2] = 6.0 * c20 * z;
-            ModuleBase::timer::end("DFPT_Pert", "grad_real_ylm");
-            return;
-        case 1:
-            grad[0] = -c2 * z;
-            grad[2] = -c2 * x;
-            ModuleBase::timer::end("DFPT_Pert", "grad_real_ylm");
-            return;
-        case 2:
-            grad[0] = 2.0 * c20 * x;
-            grad[1] = -2.0 * c20 * y;
-            ModuleBase::timer::end("DFPT_Pert", "grad_real_ylm");
-            return;
-        }
-        break;
-    default:
-        ModuleBase::WARNING_QUIT("DFPT_Pert::grad_real_ylm", "grad_real_ylm implemented for l<=2 only (DFPT NC path).");
-    }
-    ModuleBase::timer::end("DFPT_Pert", "grad_real_ylm");
-}
-
-void DFPT_Pert::build_vkb_dk(int it,
-                             int ia,
-                             int dir,
-                             const std::vector<ModuleBase::Vector3<double>>& gk,
-                             std::vector<std::vector<std::complex<double>>>& vkb,
-                             std::vector<std::vector<std::complex<double>>>& dvkb) const
-{
-    ModuleBase::TITLE("DFPT_Pert", "build_vkb_dk");
-    ModuleBase::timer::start("DFPT_Pert", "build_vkb_dk");
-    const pseudo& ncpp = ucell_->atoms[it].ncpp;
-    const int nh = ncpp.nh;
-    const int ngk = static_cast<int>(gk.size());
-    const ModuleBase::Vector3<double>& tau = ucell_->atoms[it].tau[ia];
-    if (static_cast<int>(vkb.size()) != nh || static_cast<int>(vkb[0].size()) != ngk)
-    {
-        ModuleBase::WARNING_QUIT("DFPT_Pert::build_vkb_dk", "vkb must be built on the same gk list first.");
-    }
-    dvkb.assign(nh, std::vector<std::complex<double>>(ngk, std::complex<double>(0.0, 0.0)));
-    if (nh == 0)
-    {
-        ModuleBase::timer::end("DFPT_Pert", "build_vkb_dk");
-        return;
-    }
-    const double dg = 1.0e-4; // bohr^-1, radial central-difference step
-    const double gmag_zero_tol = 1.0e-10; ///< empirical parameter: |G| floor (2*pi/lat0) for the angular terms
-    int mu = 0;
-    for (int ib = 0; ib < ncpp.nbeta; ++ib)
-    {
-        const int l = ncpp.lll[ib];
-        const std::complex<double> pref = std::pow(std::complex<double>(0.0, -1.0), l); // (-i)^l
-        for (int m = 0; m < 2 * l + 1; ++m)
-        {
-            const int mr = (m == 0) ? 0 : ((m % 2 == 1) ? (m + 1) / 2 : -(m / 2));
-            for (int ig = 0; ig < ngk; ++ig)
-            {
-                const ModuleBase::Vector3<double>& G = gk[ig];
-                const double gmag = std::sqrt(G * G);      // 2*pi/lat0 units
-                const double gnorm = gmag * ucell_->tpiba; // bohr^-1
-                const double vq0 = radial_vq(it, ib, gnorm);
-                const double dvq = (radial_vq(it, ib, gnorm + dg) - radial_vq(it, ib, std::max(0.0, gnorm - dg)))
-                                   / (dg * (gnorm > dg ? 2.0 : 1.0));
-                const double arg = -ModuleBase::TWO_PI * (G * tau);
-                const std::complex<double> phase(std::cos(arg), std::sin(arg));
-                const std::complex<double> dphase = std::complex<double>(0.0, -ModuleBase::TWO_PI * tau[dir]) * phase;
-                double dy[3] = {0.0, 0.0, 0.0};
-                double ylm = 0.0;
-                if (gmag > gmag_zero_tol)
-                {
-                    const ModuleBase::Vector3<double> ghat = G * (1.0 / gmag);
-                    ylm = real_ylm(l, mr, ghat);
-                    grad_real_ylm(l, mr, ghat, dy);
-                    const double gdir[3] = {ghat.x, ghat.y, ghat.z};
-                    // chain rule dghat/dk_dir = (e_dir - ghat*ghat_dir)/|G|
-                    double dylm_dir = 0.0;
-                    for (int c = 0; c < 3; ++c)
-                    {
-                        dylm_dir += dy[c] * ((c == dir ? 1.0 : 0.0) - gdir[c] * gdir[dir]);
-                    }
-                    dylm_dir /= gmag;
-                    // radial chain: dg/dk_dir = tpiba * ghat_dir
-                    const double dradial = dvq * ucell_->tpiba * gdir[dir];
-                    dvkb[mu][ig] = pref * phase * (dylm_dir * vq0 + ylm * dradial) + pref * ylm * vq0 * dphase;
-                }
-                else
-                {
-                    // degenerate |G| = 0: only the l = 0 channel survives
-                    // (real_ylm convention); keep only the phase term
-                    ylm = (l == 0) ? 0.5 * std::sqrt(1.0 / ModuleBase::PI) : 0.0;
-                    dvkb[mu][ig] = pref * ylm * vq0 * dphase;
-                }
-            }
-            ++mu;
-        }
-    }
-    ModuleBase::timer::end("DFPT_Pert", "build_vkb_dk");
-}
-
-void DFPT_Pert::dVnl_dtau(int atom_idx,
-                          int dir,
-                          const ModuleBase::Vector3<double>& q_cart,
-                          const psi::Psi<std::complex<double>>& psi,
-                          int k_idx,
-                          std::vector<std::vector<std::complex<double>>>& dv_psi)
-{
-    ModuleBase::TITLE("DFPT_Pert", "dVnl_dtau");
-    ModuleBase::timer::start("DFPT_Pert", "dVnl_dtau");
-    int it = 0;
-    int ia = 0;
-    atom_index(atom_idx, it, ia);
-    if (ia < 0)
-    {
-        ModuleBase::timer::end("DFPT_Pert", "dVnl_dtau");
-        return;
-    }
-    const pseudo& ncpp = ucell_->atoms[it].ncpp;
-    if (ncpp.tvanp || ncpp.has_so)
-    {
-        // the separable NC path documented in C1; ultrasoft and spin-orbit
-        // projectors are deferred (their D and augmentation have |k+q| shifts
-        // that need the USPP machinery).
-        ModuleBase::WARNING_QUIT("DFPT_Pert::dVnl_dtau",
-                                 "DFPT nonlocal first-order potential is implemented "
-                                 "for normal-conserving separable pseudopotentials only.");
-    }
-    const int nh = ncpp.nh;
-
-    // projector -> (radial beta index, m channel) table, matching build_vkb.
-    std::vector<int> mu_ib(nh, 0);
-    std::vector<int> mu_m(nh, 0);
-    int mu_idx = 0;
-    for (int ib = 0; ib < ncpp.nbeta; ++ib)
-    {
-        const int l = ncpp.lll[ib];
-        for (int m = 0; m < 2 * l + 1; ++m)
-        {
-            if (mu_idx < nh)
-            {
-                mu_ib[mu_idx] = ib;
-                mu_m[mu_idx] = m;
-            }
-            ++mu_idx;
-        }
-    }
-
-    // incoming k basis: G = k + G' (pw_wfc k-basis index)
-    const int npwk = pw_wfc_->npwk[k_idx];
-    std::vector<ModuleBase::Vector3<double>> gk_in(npwk);
-    for (int ig = 0; ig < npwk; ++ig)
-    {
-        gk_in[ig] = pw_wfc_->getgpluskcar(k_idx, ig);
-    }
-    std::vector<std::vector<std::complex<double>>> vkb_in;
-    build_vkb(it, ia, gk_in, vkb_in);
-
-    // outgoing k+q basis
-    DFPT_KQ_Basis kq;
-    kq.init(pw_wfc_, pw_rho_, q_cart, k_idx);
-    const int npwk_kq = kq.get_npwk();
-    std::vector<ModuleBase::Vector3<double>> gk_out(npwk_kq);
-    for (int igl = 0; igl < npwk_kq; ++igl)
-    {
-        gk_out[igl] = kq.get_gpluskq(igl);
-    }
-    std::vector<std::vector<std::complex<double>>> vkb_out;
-    build_vkb(it, ia, gk_out, vkb_out);
-
-    const int nbands = psi.get_nbands();
-    dv_psi.assign(nbands, std::vector<std::complex<double>>(npwk_kq, std::complex<double>(0.0, 0.0)));
-
-    for (int iband = 0; iband < nbands; ++iband)
-    {
-        // becp_nu(k) = sum_G' conj(vkb_in[nu][G']) psi(G')
-        std::vector<std::complex<double>> becp(nh, std::complex<double>(0.0, 0.0));
-        for (int nu = 0; nu < nh; ++nu)
-        {
-            for (int ig = 0; ig < npwk; ++ig)
-            {
-                becp[nu] += std::conj(vkb_in[nu][ig]) * psi(k_idx, iband, ig);
-            }
-        }
-        // dcbecp = D * becp with D_{mu,nu} = dion(ib_mu, ib_nu) delta_{m_mu, m_nu}
-        std::vector<std::complex<double>> dcbecp(nh, std::complex<double>(0.0, 0.0));
-        for (int mu = 0; mu < nh; ++mu)
-        {
-            for (int nu = 0; nu < nh; ++nu)
-            {
-                if (mu_m[mu] != mu_m[nu])
-                {
-                    continue;
-                }
-                dcbecp[mu] += ncpp.dion(mu_ib[mu], mu_ib[nu]) * becp[nu];
-            }
-        }
-        // term A: i (k+q+G'')_dir * (Vnl |psi>) on the k+q basis
-        std::vector<std::complex<double>> term_a(npwk_kq, std::complex<double>(0.0, 0.0));
-        for (int igl = 0; igl < npwk_kq; ++igl)
-        {
-            std::complex<double> vnlpsi(0.0, 0.0);
-            for (int mu = 0; mu < nh; ++mu)
-            {
-                vnlpsi += vkb_out[mu][igl] * dcbecp[mu];
-            }
-            term_a[igl] = std::complex<double>(0.0, 1.0) * (ucell_->tpiba * gk_out[igl][dir]) * vnlpsi;
-        }
-        // term B: Vnl [i (k+G')_dir |psi>]
-        std::vector<std::complex<double>> becp_dpsi(nh, std::complex<double>(0.0, 0.0));
-        for (int nu = 0; nu < nh; ++nu)
-        {
-            for (int ig = 0; ig < npwk; ++ig)
-            {
-                const std::complex<double> dpsi_ig
-                    = std::complex<double>(0.0, 1.0) * (ucell_->tpiba * gk_in[ig][dir]) * psi(k_idx, iband, ig);
-                becp_dpsi[nu] += std::conj(vkb_in[nu][ig]) * dpsi_ig;
-            }
-        }
-        std::vector<std::complex<double>> dcbecp_dpsi(nh, std::complex<double>(0.0, 0.0));
-        for (int mu = 0; mu < nh; ++mu)
-        {
-            for (int nu = 0; nu < nh; ++nu)
-            {
-                if (mu_m[mu] != mu_m[nu])
-                {
-                    continue;
-                }
-                dcbecp_dpsi[mu] += ncpp.dion(mu_ib[mu], mu_ib[nu]) * becp_dpsi[nu];
-            }
-        }
-        std::vector<std::complex<double>> term_b(npwk_kq, std::complex<double>(0.0, 0.0));
-        for (int igl = 0; igl < npwk_kq; ++igl)
-        {
-            std::complex<double> vnl_dpsi(0.0, 0.0);
-            for (int mu = 0; mu < nh; ++mu)
-            {
-                vnl_dpsi += vkb_out[mu][igl] * dcbecp_dpsi[mu];
-            }
-            term_b[igl] = vnl_dpsi;
-        }
-        for (int igl = 0; igl < npwk_kq; ++igl)
-        {
-            // GS exp(-2pi gk.tau) projector convention: dVnl/dtau_dir
-            // |psi> = -i (k+q+G'')_dir (Vnl|psi>) + Vnl[i (k+G')_dir |psi>]
-            dv_psi[iband][igl] = term_b[igl] - term_a[igl];
-        }
-    }
-    ModuleBase::timer::end("DFPT_Pert", "dVnl_dtau");
-}
-
 void DFPT_Pert::build_dv_u(int q_idx, int atom_idx, int dir, DFPT_PW_Data& data)
 {
     ModuleBase::TITLE("DFPT_Pert", "build_dv_u");
@@ -865,141 +420,6 @@ void DFPT_Pert::d2vloc_r(int atom_idx, int da, int db, std::vector<std::complex<
     dv2_r.assign(pw_rho_->nrxx, std::complex<double>(0.0, 0.0));
     pw_rho_->recip2real(dv2_recip.data(), dv2_r.data());
     ModuleBase::timer::end("DFPT_Pert", "d2vloc_r");
-}
-
-void DFPT_Pert::apply_d2vnl(int atom_idx,
-                            int da,
-                            int db,
-                            const ModuleBase::Vector3<double>& q_eff,
-                            bool include_middle,
-                            const psi::Psi<std::complex<double>>& psi,
-                            int k_idx,
-                            std::vector<std::vector<std::complex<double>>>& d2v_psi) const
-{
-    ModuleBase::TITLE("DFPT_Pert", "apply_d2vnl");
-    ModuleBase::timer::start("DFPT_Pert", "apply_d2vnl");
-    int it = 0;
-    int ia = 0;
-    atom_index(atom_idx, it, ia);
-    if (ia < 0)
-    {
-        ModuleBase::timer::end("DFPT_Pert", "apply_d2vnl");
-        return;
-    }
-    const pseudo& ncpp = ucell_->atoms[it].ncpp;
-    if (ncpp.tvanp || ncpp.has_so)
-    {
-        ModuleBase::WARNING_QUIT("DFPT_Pert::apply_d2vnl",
-                                 "DFPT second-order nonlocal potential is implemented "
-                                 "for normal-conserving separable pseudopotentials only.");
-    }
-    const int nh = ncpp.nh;
-    const int nbands = psi.get_nbands();
-
-    // projector -> (radial index, m channel) table, matching build_vkb
-    std::vector<int> mu_ib(nh, 0);
-    std::vector<int> mu_m(nh, 0);
-    int mu_idx = 0;
-    for (int ib = 0; ib < ncpp.nbeta; ++ib)
-    {
-        const int l = ncpp.lll[ib];
-        for (int m = 0; m < 2 * l + 1; ++m)
-        {
-            if (mu_idx < nh)
-            {
-                mu_ib[mu_idx] = ib;
-                mu_m[mu_idx] = m;
-            }
-            ++mu_idx;
-        }
-    }
-
-    // incoming k basis and outgoing k+q basis projectors (same atom)
-    const int npwk = pw_wfc_->npwk[k_idx];
-    std::vector<ModuleBase::Vector3<double>> gk_in(npwk);
-    for (int ig = 0; ig < npwk; ++ig)
-    {
-        gk_in[ig] = pw_wfc_->getgpluskcar(k_idx, ig);
-    }
-    std::vector<std::vector<std::complex<double>>> vkb_in;
-    build_vkb(it, ia, gk_in, vkb_in);
-    DFPT_KQ_Basis kq;
-    kq.init(pw_wfc_, pw_rho_, q_eff, k_idx);
-    const int npwk_kq = kq.get_npwk();
-    std::vector<ModuleBase::Vector3<double>> gk_out(npwk_kq);
-    for (int igl = 0; igl < npwk_kq; ++igl)
-    {
-        gk_out[igl] = kq.get_gpluskq(igl);
-    }
-    std::vector<std::vector<std::complex<double>>> vkb_out;
-    build_vkb(it, ia, gk_out, vkb_out);
-
-    d2v_psi.assign(nbands, std::vector<std::complex<double>>(npwk_kq, std::complex<double>(0.0, 0.0)));
-    for (int iband = 0; iband < nbands; ++iband)
-    {
-        // becp and its (k+G')-weighted variants: becp_x = sum x(G') |beta><psi|
-        std::vector<std::complex<double>> becp(nh, std::complex<double>(0.0, 0.0));
-        std::vector<std::complex<double>> becp_a(nh, std::complex<double>(0.0, 0.0));
-        std::vector<std::complex<double>> becp_b(nh, std::complex<double>(0.0, 0.0));
-        std::vector<std::complex<double>> becp_ab(nh, std::complex<double>(0.0, 0.0));
-        for (int nu = 0; nu < nh; ++nu)
-        {
-            for (int ig = 0; ig < npwk; ++ig)
-            {
-                const std::complex<double> vc = std::conj(vkb_in[nu][ig]) * psi(k_idx, iband, ig);
-                const double kp_da = ucell_->tpiba * gk_in[ig][da];
-                const double kp_db = ucell_->tpiba * gk_in[ig][db];
-                becp[nu] += vc;
-                becp_a[nu] += kp_da * vc;
-                becp_b[nu] += kp_db * vc;
-                becp_ab[nu] += kp_da * kp_db * vc;
-            }
-        }
-        // D contraction with the same-m selection rule as dVnl_dtau
-        std::vector<std::complex<double>> d0(nh, std::complex<double>(0.0, 0.0));
-        std::vector<std::complex<double>> da_(nh, std::complex<double>(0.0, 0.0));
-        std::vector<std::complex<double>> db_(nh, std::complex<double>(0.0, 0.0));
-        std::vector<std::complex<double>> dab(nh, std::complex<double>(0.0, 0.0));
-        for (int mu = 0; mu < nh; ++mu)
-        {
-            for (int nu = 0; nu < nh; ++nu)
-            {
-                if (mu_m[mu] != mu_m[nu])
-                {
-                    continue;
-                }
-                const double dij = ncpp.dion(mu_ib[mu], mu_ib[nu]);
-                d0[mu] += dij * becp[nu];
-                da_[mu] += dij * becp_a[nu];
-                db_[mu] += dij * becp_b[nu];
-                dab[mu] += dij * becp_ab[nu];
-            }
-        }
-        // chi(G'') = sum_mu vkb_out,mu [ -kq_da kq_db d0 - dab
-        //                              + (include_middle ? kq_da db_ + kq_db da_ : 0) ]_mu
-        // QE ground truth (dynmat_us.f90 + phq_init.f90): the KB second-order
-        // term pairs gammap (integer-G (k+G)_da(k+G)_db derivative of beta)
-        // with becp1 = <beta_k|psi_k> and the same-atom alphap_a* alphap_b
-        // middle product; everything is built at k with integer-G momentum
-        // factors, so the caller passes q_eff = 0 and the kernel is
-        // q-independent for every q.
-        for (int igl = 0; igl < npwk_kq; ++igl)
-        {
-            const double kq_da = ucell_->tpiba * gk_out[igl][da];
-            const double kq_db = ucell_->tpiba * gk_out[igl][db];
-            std::complex<double> chi(0.0, 0.0);
-            for (int mu = 0; mu < nh; ++mu)
-            {
-                chi += vkb_out[mu][igl] * (-kq_da * kq_db * d0[mu] - dab[mu]);
-                if (include_middle)
-                {
-                    chi += vkb_out[mu][igl] * (kq_da * db_[mu] + kq_db * da_[mu]);
-                }
-            }
-            d2v_psi[iband][igl] = chi;
-        }
-    }
-    ModuleBase::timer::end("DFPT_Pert", "apply_d2vnl");
 }
 
 void DFPT_Pert::build_efield(const ModuleBase::Vector3<double>& field, DFPT_PW_Data& data)
