+++ source/source_pw/module_dfpt/test/CMakeLists.txt
@@ -40,6 +40,8 @@ AddTest(
 	../dfpt_pw.cpp
 	../dfpt_pw_data.cpp
 	../dfpt_pert.cpp
+	../dfpt_pert_vkb.cpp
+	../dfpt_pert_nl.cpp
 	../dfpt_kq_basis.cpp
 	../dfpt_stern.cpp
 	../dfpt_rho.cpp
